Added tooltips and ui change
basically added tooltips and made the ui nicer.

+++ Workplace/Assets/Scripts/PlayerData.cs
@@ -10,7 +10,7 @@ public class PlayerData : MonoBehaviour
     public int Hunger;
     public int Happiness;
     public int Day;
-    public int Intelligence;
+    public float Intelligence;
     public string Name = "PlayerNameNotFound";
     private void Awake()
     {
